+++ packages/evokit-picture/CHANGELOG.md
@@ -1,8 +1,11 @@
+[README]: /packages/evokit-picture/README.md
+
 [create_theme]: /docs/base/theme.md
 
 # EvoKit - Picture
 
 [![](https://img.shields.io/npm/v/evokit-picture.svg)](https://www.npmjs.com/package/evokit-picture)
+[![](https://img.shields.io/badge/page-README-42b983)][README]
 
 ---